requirements.txt
fastapi==0.107.0
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
alembic==1.13.0
python-dotenv==1.0.0
email-validator==2.1.0
redis==4.6.0
requests==2.32.5
geoalchemy2==0.18.0
httpx==0.27.0
openpyxl==3.1.5
mangum==0.17.0 